Update sdv 1.6 smapi 4.5

+++ Framework/TimeHelper.cs
@@ -32,7 +32,7 @@ namespace ItsStardewTime.Framework
         /// <summary>The percentage of the <see cref="CurrentDefaultTickInterval"/> that's elapsed since the last tick.</summary>
         public double TickProgress
         {
-            get => (double)TickProgressBackingStore / CurrentDefaultTickInterval;
+            get => TickProgressBackingStore / CurrentDefaultTickInterval;
             set => TickProgressBackingStore = value * CurrentDefaultTickInterval;
         }
